+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 * **Dropped support for Python 2** because it's a pain to support and was
   causing bugs with handling unicode.
+* Remove hacky `login_2fa` command, use `login_browser` instead
 
 **0.15.1 (2017-12-12)**